requirements.txt
python-dotenv 
azure-identity
azure-ai-projects
semantic-kernel